# internal/server/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Corvo is a distributed background job queue. All endpoints are under `/api/v1`. Authentication via `X-API-Key` header or `Authorization: Bearer` token.",
        "title": "Corvo API",
        "contact": {
            "url": "https://corvohq.com"
        },
        "version": "v1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/search/fulltext": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Searches job payloads, tags, and queue names using SQLite FTS or LIKE fallback.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Full-text search jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default: 50, max: 500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/usage/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns aggregate token and cost data for AI/agent jobs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budgets"
                ],
                "summary": "Token and cost usage summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time period (e.g. 24h, 168h)",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Group by field",
                        "name": "group_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "List webhooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.webhookConfig"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Create or update a webhook",
                "parameters": [
                    {
                        "description": "Webhook config",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.webhookConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/{queue}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Enqueues a job using the raw HTTP body as payload. Useful for receiving webhooks from third-party services.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Enqueue via webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue name",
                        "name": "queue",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Priority: low, normal, high, critical",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Deduplication key",
                        "name": "unique_key",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max retry attempts",
                        "name": "max_retries",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 schedule time",
                        "name": "scheduled_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tags as comma-separated key:value pairs",
                        "name": "tags",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate exists (unique_key)",
                        "schema": {
                            "$ref": "#/definitions/store.EnqueueResult"
                        }
                    },
                    "201": {
                        "description": "Job enqueued",
                        "schema": {
                            "$ref": "#/definitions/store.EnqueueResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "server.webhookConfig": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "last_delivery_at": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "last_status_code": {
                    "type": "integer"
                },
                "retry_limit": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "store.EnqueueResult": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unique_existing": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer token: `Authorization: Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}